Reducing number of line
