+++ TelCo.ColorCoder/SubModule/ColorPair.cs
@@ -2,6 +2,7 @@
 using System.Drawing;
 using System.Text;
 using System.Collections.Generic;
+using static TelCo.ColorCoder.SubModule.ColorDefination;
 
 namespace TelCo.ColorCoder.SubModule
 {
@@ -12,6 +13,30 @@ namespace TelCo.ColorCoder.SubModule
         public override string ToString()
         {
             return string.Format("MajorColor:{0}, MinorColor:{1}", majorColor.Name, minorColor.Name);
+        }
+
+        public static ColorPair GetColorFromPairNumber(int pairNumber)
+        {
+            // The function supports only 1 based index. Pair numbers valid are from 1 to 25
+            int minorSize = colorMapMinor.Length;
+            int majorSize = colorMapMajor.Length;
+            if (pairNumber < 1 || pairNumber > minorSize * majorSize)
+            {
+                throw new ArgumentOutOfRangeException(
+                    string.Format("Argument PairNumber:{0} is outside the allowed range", pairNumber));
+            }
+            // Find index of major and minor color from pair number
+            int zeroBasedPairNumber = pairNumber - 1;
+            int majorIndex = zeroBasedPairNumber / minorSize;
+            int minorIndex = zeroBasedPairNumber % minorSize;
+            // Construct the return val from the arrays
+            ColorPair pair = new ColorPair()
+            {
+                majorColor = colorMapMajor[majorIndex],
+                minorColor = colorMapMinor[minorIndex]
+            };
+            // return the value
+            return pair;
         }
     }
 }
